Fixed warnings

+++ includes/PerspectiveCamera.h
@@ -5,7 +5,7 @@
 
 class PerspectiveCamera : public Camera{ 
 public:
-	PerspectiveCamera(float3 position, float3 lookAt, float3 up, int fov, float ratio, float nearPlane, float farPlane);
+	PerspectiveCamera(float3 position, float3 lookAt, float3 up, float fov, float ratio, float nearPlane, float farPlane);
 	~PerspectiveCamera();
 
 	float4x4 getViewMatrix();
